fix: move hooks
