+++ packages/hierarchy-browser/src/CommunityCard/AdjacentCommunities/AdjacentCommunities.tsx
@@ -5,16 +5,16 @@
 import { Spinner } from '@fluentui/react'
 import React, { memo } from 'react'
 import styled from 'styled-components'
+import { IEntityDetail } from '../..'
 import CommunityEdgeList from '../../NeighborList/CommunityEdgeList'
 import { ScrollArea } from '../../ScollArea'
 import { CommunityDataProvider } from '../../common/dataProviders'
 import { ICardFontStyles, useThemesAccentStyle } from '../../hooks/theme'
-import { useAdjacentCommunityData } from '../../hooks/useAdjacentCommunityData'
-import { useEdgeSelection } from '../../hooks/useEdgeSelection'
-import { useExpandedPanel } from '../../hooks/useExpandedPanel'
-import { IEntityDetail } from '../../types'
 import { CommunityTable } from '../CommunityTable'
 import { TableExpander } from '../TableExpander'
+import { useAdjacentCommunityData } from './hooks/useAdjacentCommunityData'
+import { useEdgeSelection } from './hooks/useEdgeSelection'
+import { useExpandedPanel } from './hooks/useExpandedPanel'
 
 const ENTITY_LOADER_MSG = 'Fetching entity data...'
